services/shared/pyproject.toml
[project]
name = "farmer-code-contracts"
version = "0.1.0"
description = "Shared contracts and utilities for Farmer Code services"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contracts", "src/github_integration", "src/worktree_manager"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "Q", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
